Add tests for bug in OneToOne document deletions
Hibernate Search version: 7.2.4.Final
Hibernate Orm version: 6.6.34.Final

This adds tests which reproduce a bug related to deleting
documents of bidirectional OneToOne relations. If we have a
bidirectional OneToOne relation defined as follows:

  @Entity
  public class Parent {

      @Id
      private Long id;

      @FullTextField
      @Column(name = "PARENT_NAME")
      private String parentName;

      @OneToOne(
        mappedBy = "parent",
        cascade = CascadeType.ALL)
      private Child child;
  }

--

  @Entity
  @Indexed
  public class Child {

      @Id
      @DocumentId
      private Long id;

      @FullTextField
      @Column(name = "CHILD_NAME")
      private String childName;

      @MapsId
      @JoinColumn(name = "P_ID")
      @OneToOne
      @IndexedEmbedded
      @AssociationInverseSide(
        inversePath =
          @ObjectPath(
            @PropertyValue(
              propertyName = "child")))
      private Parent parent;
  }

Then, if we try to delete the parent the child will get
deleted from the database, but the corresponding document
will remain in the search index. Also, if the child entity
contains a lazy-loaded property, the transaction will fail.

Based on some debugging, it looks as if when hibernate
is processing the action queue on transaction commit, the
child gets deleted first and the indexing plan correctly
schedules a deletion of the document by setting the current
EntityStatus to ABSENT, but when the parent gets deleted
the current EntityStatus is changed back to PRESENT which
instead schedules an addOrUpdate operation for the already
detached entity.

This problem can be resolved by removing the @JoinColumn
annotation from the Child.parent field, but this could
require renaming of database columns.

Also reproducible on:
Hibernate Search version: 8.1.2.Final
Hibernate Orm version: 7.1.2.Final

Not reproducible on:
Hibernate Search version: 7.1.2.Final
Hibernate Orm version: 6.5.3.Final